return CancellationCode from killMutation [#CLICKHOUSE-3912]
